feat: http request to cbr

+++ MiniBank.Core/Interfaces/ICurrencyRateConversionService.cs
@@ -2,6 +2,6 @@
 {
     public interface ICurrencyRateConversionService
     {
-        int ConvertCurrencyRate(int amount, string fromCurrencyCode, string toCurrencyCode);
+        double ConvertCurrencyRate(double amount, string fromCurrencyCode, string toCurrencyCode);
     }
 }
